// front/open-api-generator/client/models/FindIssueModel.ts
/**
 * Roadmap for Jira API
 * Endpoint mapping for Roadmap for Jira App.
 *
 * OpenAPI spec version: 2.0
 * Contact: contact@mrrafael.ca
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class FindIssueModel {
    'guid': string;
    'filter': string;
    'subTasks': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "guid",
            "baseName": "guid",
            "type": "string",
            "format": ""
        },
        {
            "name": "filter",
            "baseName": "filter",
            "type": "string",
            "format": ""
        },
        {
            "name": "subTasks",
            "baseName": "subTasks",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return FindIssueModel.attributeTypeMap;
    }

    public constructor() {
    }
}